% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/persistence-diagram.R
\name{as_diagram.ttk_persistence_diagram}
\alias{as_diagram.ttk_persistence_diagram}
\title{Convert ttk_persistence_diagram to TDA diagram format}
\usage{
as_diagram.ttk_persistence_diagram(x, list = TRUE, ...)
}
\arguments{
\item{x}{A ttk_persistence_diagram object}

\item{list}{Logical, whether to wrap in a list (default: TRUE for TDA compatibility)}

\item{...}{Additional arguments (currently unused)}
}
\value{
An object of class 'diagram' compatible with TDA package
}
\description{
Convert ttk_persistence_diagram to TDA diagram format
}
