% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/persistence-diagram.R
\name{plot.ttk_persistence_diagram}
\alias{plot.ttk_persistence_diagram}
\title{Plot method for persistence diagram}
\usage{
\method{plot}{ttk_persistence_diagram}(x, ...)
}
\arguments{
\item{x}{A ttk_persistence_diagram object}

\item{...}{Additional arguments passed to plot}
}
\description{
Plot method for persistence diagram
}
