% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contour-tree.R
\name{ttk_contour_tree}
\alias{ttk_contour_tree}
\title{Compute contour tree}
\usage{
ttk_contour_tree(mesh, vertex_scalar_field = "data", simplify_tree = FALSE)
}
\arguments{
\item{mesh}{A mesh3d object with scalar field}

\item{vertex_scalar_field}{Name of the scalar field}

\item{simplify_tree}{Logical, simplify the tree (default: FALSE)}
}
\value{
A list with nodes, edges, and hierarchy information
}
\description{
Compute contour tree
}
