% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/critical-points.R
\name{ttk_critical_points}
\alias{ttk_critical_points}
\title{Extract critical points from a scalar field on a mesh}
\usage{
ttk_critical_points(
  mesh,
  vertex_scalar_field = "data",
  compute_minima = TRUE,
  compute_maxima = TRUE,
  compute_saddle_points = TRUE
)
}
\arguments{
\item{mesh}{A mesh3d object with a scalar field}

\item{vertex_scalar_field}{Name of the scalar field (default: "data")}

\item{compute_minima}{Logical, compute minima (default: TRUE)}

\item{compute_maxima}{Logical, compute maxima (default: TRUE)}

\item{compute_saddle_points}{Logical, compute saddle points (default: TRUE)}
}
\value{
A list with critical points coordinates and types
}
\description{
Extract critical points from a scalar field on a mesh
}
