% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge-tree.R
\name{ttk_merge_tree}
\alias{ttk_merge_tree}
\title{Compute merge tree}
\usage{
ttk_merge_tree(
  mesh,
  vertex_scalar_field = "data",
  tree_type = "join",
  simplify_tree = FALSE
)
}
\arguments{
\item{mesh}{A mesh3d object with scalar field}

\item{vertex_scalar_field}{Name of the scalar field}

\item{tree_type}{Type of merge tree ("join" or "split", default: "join")}

\item{simplify_tree}{Logical, simplify the tree (default: FALSE)}
}
\value{
A list with tree structure information
}
\description{
Compute merge tree
}
