% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/morse-smale-complex.R
\name{ttk_morse_smale_complex}
\alias{ttk_morse_smale_complex}
\title{Compute Morse-Smale complex}
\usage{
ttk_morse_smale_complex(
  mesh,
  vertex_scalar_field = "data",
  simplify_threshold = 0
)
}
\arguments{
\item{mesh}{A mesh3d object with scalar field}

\item{vertex_scalar_field}{Name of the scalar field}

\item{simplify_threshold}{Simplification threshold (default: 0.0)}
}
\value{
A list with cells, separatrices, and critical points
}
\description{
Compute Morse-Smale complex
}
