% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/persistence-diagram.R
\name{ttk_persistence_diagram}
\alias{ttk_persistence_diagram}
\title{Compute persistence diagram}
\usage{
ttk_persistence_diagram(
  mesh,
  vertex_scalar_field = "data",
  simplify_saddles = FALSE
)
}
\arguments{
\item{mesh}{A mesh3d object with scalar field}

\item{vertex_scalar_field}{Name of the scalar field}

\item{simplify_saddles}{Logical, simplify saddle pairs (default: FALSE)}
}
\value{
A data frame with columns: birth, death, dimension, persistence
}
\description{
Compute persistence diagram
}
